Improve tests structure
