+++ pyproject.toml
@@ -62,3 +62,14 @@ flake8 = "^4.0.1"
 
 [tool.poetry.urls]
 "Bug Tracker" = "https://github.com/dsoftwareinc/fakeredis-py/issues"
+
+[tool.pytest.ini_options]
+markers = [
+    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
+    "fake: run tests only with fake redis",
+    "real: run tests with a locally running real Redis server",
+    "disconnected",
+    "min_server",
+    "decode_responses",
+]
+asyncio_mode="strict"
