--- pytest.ini
@@ -1,8 +1,0 @@
-[pytest]
-markers =
-    slow: marks tests as slow (deselect with '-m "not slow"')
-    fake: run tests only with fake redis
-    real: run tests with a locally running real Redis server
-    redis4_and_above: run tests for redis-py 4.1.2 and above
-    redis3_and_above: run tests for redis-py 3.x and above
-    @extra.run_test_if_redis_ver('below','3'): run tests for redis-py 2.x and below
