+++ test/test_lua.py
@@ -15,6 +15,12 @@ import testtools
 
 lupa = pytest.importorskip("lupa")
 
+fake_only = pytest.mark.parametrize(
+    'create_redis',
+    [pytest.param('FakeStrictRedis', marks=pytest.mark.fake)],
+    indirect=True
+)
+
 
 @pytest_asyncio.fixture
 def r(request, create_redis):
@@ -471,7 +477,7 @@ def test_script(r):
     assert result == b'42'
 
 
-@testtools.fake_only
+@fake_only
 def test_lua_log(r, caplog):
     logger = fakeredis._server.LOGGER
     script = """
@@ -498,7 +504,7 @@ def test_lua_log_no_message(r):
         script()
 
 
-@testtools.fake_only
+@fake_only
 def test_lua_log_different_types(r, caplog):
     logger = fakeredis._server.LOGGER
     script = "redis.log(redis.LOG_DEBUG, 'string', 1, true, 3.14, 'string')"
@@ -517,7 +523,7 @@ def test_lua_log_wrong_level(r):
         script()
 
 
-@testtools.fake_only
+@fake_only
 def test_lua_log_defined_vars(r, caplog):
     logger = fakeredis._server.LOGGER
     script = """
