+++ test/testtools.py
@@ -6,6 +6,7 @@ from packaging.version import Version
 
 REDIS_VERSION = Version(redis.__version__)
 
+
 # Wrap some redis commands to abstract differences between redis-py 2 and 3.
 def zadd(r, key, d, *args, **kwargs):
     if REDIS_VERSION >= Version('3'):
@@ -21,12 +22,6 @@ def run_test_if_redis_ver(condition: str, ver: str):
         reason=f"Test is only applicable to redis-py {ver} and above"
     )
 
-
-fake_only = pytest.mark.parametrize(
-    'create_redis',
-    [pytest.param('FakeStrictRedis', marks=pytest.mark.fake)],
-    indirect=True
-)
 
 _lua_module = importlib.util.find_spec("lupa")
 run_test_if_lupa = pytest.mark.skipif(
